requirements.txt
#requirements.txt
openai
pyperclip
python-dotenv
argparse
colorama